refactor: consolidate byte-identical helper pairs onto owner modules
