+++ src/agents/tools-effective-inventory-build.ts
@@ -107,7 +107,7 @@ function readMatchingTool(
 
 // Raw tool arrays can contain getters/proxies from plugin boundaries. Read
 // defensively; projection diagnostics handle the exact unreadable entry later.
-function buildReadableRawToolsByName(
+export function buildReadableToolsByName(
   tools: readonly AnyAgentTool[],
 ): ReadonlyMap<string, AnyAgentTool> {
   const toolsByName = new Map<string, AnyAgentTool>();
@@ -185,7 +185,7 @@ export function buildRuntimeCompatibleToolInventory(params: {
   entries: EffectiveToolInventoryEntry[];
   notices: EffectiveToolInventoryNotice[];
 } {
-  const rawToolsByName = buildReadableRawToolsByName(params.tools);
+  const rawToolsByName = buildReadableToolsByName(params.tools);
   const preNormalizationProjection = filterProviderNormalizableTools(params.tools);
   const preNormalizationDiagnostics: RuntimeToolSchemaDiagnostic[] = [
     ...preNormalizationProjection.diagnostics,
