+++ src/plugins/cli-gateway-nodes-runtime.ts
@@ -6,7 +6,7 @@ import {
   GATEWAY_CLIENT_NAMES,
 } from "../../packages/gateway-protocol/src/client-info.js";
 import { callGateway } from "../gateway/call.js";
-import { isOperatorScope, type OperatorScope } from "../gateway/operator-scopes.js";
+import { normalizeOperatorScopeList } from "../gateway/operator-scopes.js";
 import { getPluginRuntimeGatewayRequestScope } from "./runtime/gateway-request-scope.js";
 import type { PluginRuntime } from "./runtime/types.js";
 
@@ -19,15 +19,6 @@ export function resolvePluginCliNodeInvokeGatewayTimeoutMs(
     : undefined;
 }
 
-function normalizeRuntimeNodeInvokeScopes(
-  scopes: string[] | undefined,
-): OperatorScope[] | undefined {
-  if (!Array.isArray(scopes)) {
-    return undefined;
-  }
-  return scopes.filter(isOperatorScope);
-}
-
 function canPluginCliRuntimeRequestScopes(): boolean {
   const scope = getPluginRuntimeGatewayRequestScope();
   return Boolean(
@@ -37,7 +28,7 @@ function canPluginCliRuntimeRequestScopes(): boolean {
 }
 
 function resolvePluginCliRuntimeNodeInvokeScopes(scopes: string[] | undefined) {
-  const normalizedScopes = normalizeRuntimeNodeInvokeScopes(scopes);
+  const normalizedScopes = normalizeOperatorScopeList(scopes);
   return normalizedScopes && canPluginCliRuntimeRequestScopes() ? normalizedScopes : undefined;
 }
